+++ changes.md
@@ -1,5 +1,11 @@
 # Zealot
 
+## Release v0.2.0
+LTS: 12/12/2024
+
+ - Creating new sessions with a collection of predefined headers 
+ - Added HttpRequest.removeHeader
+
 ## Release v0.1.5
 LTS: 11/12/2024